@@ -1,54 +1,71 @@
-import React from 'react';
-import { Modal } from '@material-ui/core';
-import { modalStyles } from '../styles/style';
+import Backdrop from "@mui/material/Backdrop"
+import Box from "@mui/material/Box"
+import Button from "@mui/material/Button"
+import Fade from "@mui/material/Fade"
+import Modal from "@mui/material/Modal"
+import Typography from "@mui/material/Typography"
 
-type ModalPropType = {
-  open: boolean;
-  message: string;
-  handleOnClick?: () => void;
-  buttonText?: string;
-  isButton?: boolean;
-  isCenter?: boolean;
-};
+const boxStyle = {
+  position: "absolute",
+  top: "50%",
+  left: "50%",
+  transform: "translate(-50%, -50%)",
+  maxWidth: 400,
+  bgcolor: "background.paper",
+  border: "2px solid #000",
+  boxShadow: 24,
+  p: 4,
+}
 
-const CustomModal = ({
+interface CustomModalProps {
+  open: boolean
+  message: string
+  handleOnClick?: () => void
+  buttonText?: string
+  showButton?: boolean
+}
+
+function CustomModal({
   open,
   message,
   handleOnClick,
   buttonText,
-  isButton = true,
-  isCenter = true,
-}: ModalPropType) => {
-  const modalClasses = modalStyles();
-
+  showButton,
+}: CustomModalProps) {
   return (
     <Modal
+      aria-labelledby="transition-modal-title"
+      aria-describedby="transition-modal-description"
       open={open}
-      aria-labelledby="simple-modal-title"
-      aria-describedby="simple-modal-description"
+      slots={{ backdrop: Backdrop }}
+      slotProps={{ backdrop: { timeout: 500 } }}
     >
-      <div
-        className={`${modalClasses.container} ${
-          isCenter ? modalClasses.centerAll : modalClasses.centerHorizontal
-        }`}
-      >
-        <div className={modalClasses.paper}>
-          <h2 className={modalClasses.gap}>{message}</h2>
-          {isButton && (
-            <button
-              className="btn"
-              type="button"
-              onClick={() => {
-                if (handleOnClick) handleOnClick();
-              }}
+      <Fade in={open}>
+        <Box textAlign="center">
+          <Box sx={boxStyle}>
+            <Typography
+              id="transition-modal-title"
+              variant="h6"
+              component="h2"
+              sx={{ mb: 2 }}
             >
-              {buttonText}
-            </button>
-          )}
-        </div>
-      </div>
+              {message}
+            </Typography>
+            {showButton && (
+              <Button
+                onClick={() => {
+                  handleOnClick ? handleOnClick() : ""
+                }}
+                variant="contained"
+              >
+                {buttonText}
+              </Button>
+            )}
+          </Box>
+        </Box>
+      </Fade>
     </Modal>
-  );
-};
+  )
+}
 
-export default CustomModal;
+export default CustomModal
