@@ -1,8 +1,23 @@
-import React from 'react';
-import { TEXT } from '../constants/finalConstant';
+import { useTheme } from "@mui/material/styles"
+import Typography from "@mui/material/Typography"
+
+import { TEXT } from "@constants/text.const"
 
 function NotFound() {
-  return <h1>{TEXT.pageNotFound}</h1>;
+  const theme = useTheme()
+  return (
+    <Typography
+      sx={{
+        display: "flex",
+        justifyContent: "center",
+        alignContent: "center",
+        height: "100vh",
+        marginTop: theme.spacing(5),
+      }}
+    >
+      {TEXT.pageNotFound}
+    </Typography>
+  )
 }
 
-export default NotFound;
+export default NotFound
